refactor client actions into their own components
rm unnecessary parameters to service calls
