+++ src/lib/components/workflow/client-actions/batch-operation-confirmation-modal.svelte
